Move Flame Histogram class back
